+++ src/repository/OrderRepository.java
@@ -8,20 +8,20 @@ import java.util.List;
 import java.util.Map;
 
 public interface OrderRepository {
-    int getTotalSales();
+    long getTotalSales();
 
-    Map<String, Integer> getSalesByCategory();
+    Map<String, Long> getSalesByCategory();
 
     List<String> getTopSellingMenus();
 
-    Map<String, Integer> getDailySales();
+    Map<String, Long> getDailySales();
 
-    Map<String, Integer> getSalesByPeriod(String format);
+    Map<String, Long> getSalesByPeriod(String format);
 
     // 260313 [feat]: 매출 통계 고도화 신규 메서드
     Map<String, Object> getSalesStatsByPeriod(String startDate, String endDate);
-    Map<Integer, Integer> getHourlySales();
-    Map<String, Integer> getDayOfWeekSales();
+    Map<Integer, Long> getHourlySales();
+    Map<String, Long> getDayOfWeekSales();
     List<Map<String, Object>> getTopSpenders(int limit);
 
     boolean cancelOrder(long orderId);
